Better implementation of commands.
More clean imports.

better implementation of command handler.

Cleaner layout.

Tokenization support for all levels.

Major bug fixes.
