--- src/core/modules/api.ts
@@ -1,80 +1,0 @@
-import { fastify, instance } from '../../index'
-import { FastifyRequest } from 'fastify'
-
-import mineflayer from 'mineflayer'
-
-/**
- * @param {mineflayer.Bot} bot
- */
-
-declare module 'mineflayer' {
-  interface BotEvents {
-    'chat:come': (options: string[]) => Promise<void> | void
-  }
-}
-
-module.exports = (bot: mineflayer.Bot) => {
-  fastify.get(
-    '/behavior',
-    async function (
-      req: FastifyRequest<{
-        Querystring: {
-          type: 'mine' | 'come'
-          block: string | undefined
-        }
-      }>,
-      rep,
-    ) {
-      instance.logger.info(instance.strings.msg_api_received)
-
-      const command = req.query.type
-
-      if (command === 'mine') {
-        if (instance.getStates().isMining) {
-          rep.send({
-            statusCode: 'failed',
-            code: 1,
-            message: instance.strings.msg_miner_already_started,
-          })
-          return
-        }
-
-        instance.getStates().block = req.query.block
-
-        instance.goMine(req.query.block)
-        rep.send({ statusCode: 'success', code: 0, message: 'mining ' + req.query.block })
-        instance.reportToDiscord('Console', 'mine', req.query.block)
-        return
-      }
-      const options: string[] = [
-        instance.getSettings().operators[0] ? instance.getSettings().operators[0] : '',
-      ]
-      bot.emit(`chat:${command}`, options)
-      instance.reportToDiscord('Console', command)
-      rep.send({ statusCode: 'success', code: 0, message: instance.strings.msg_api_sent })
-    },
-  )
-
-  fastify.get('/health', async function (_req, res) {
-    instance.reportToDiscord('Console', 'health')
-    res.send({
-      bot: {
-        api: 'localhost:' + ':' + instance.getSettings().interfacePort,
-        inventory: 'localhost' + ':' + instance.getSettings().inventoryPort,
-        viewer: 'localhost' + ':' + instance.getSettings().viewPort,
-      },
-      server: {
-        host: bot._client.socket.address() + ':' + bot._client.socket.remotePort,
-        version: bot.version,
-      },
-    })
-  })
-
-  fastify.listen({ port: instance.getSettings().interfacePort }, async function (err, address) {
-    if (err) {
-      console.error(err)
-      process.exit(1)
-    }
-    instance.logger.info(instance.strings.msg_api_success, address)
-  })
-}
