--- src/core/modules/come.ts
@@ -1,34 +1,0 @@
-import mineflayer, { BotEvents } from 'mineflayer'
-
-import { goals } from 'mineflayer-pathfinder'
-import { instance } from '../../index'
-
-/**
- * @param {mineflayer.Bot} bot
- */
-
-module.exports = (bot: mineflayer.Bot) => {
-  bot.addChatPattern(
-    'come',
-    new RegExp(`<(.*)> ${instance.getSettings().commandsPrefix}(come|come to me)`),
-    { repeat: true, parse: true },
-  )
-
-  bot.on('chat:come' as keyof BotEvents, ([[username]]: [string]): void => {
-    if (!instance.getSettings().commandsEnabled) return
-
-    if (!instance.getSettings().operators.includes(username)) return
-    instance.reportToDiscord(username, 'come')
-
-    const player = bot.players[username]
-
-    if (!player) {
-      instance.logger.error(instance.strings.msg_finder_player_failure)
-      return
-    }
-
-    const pos = player.entity.position
-
-    bot.pathfinder.setGoal(new goals.GoalNear(pos.x, pos.y, pos.z, 2))
-  })
-}
